+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 # Changelog
 
+## 0.8.0
+
+### New features
+
+- New parsers `find_byte` and `find_byte_keep` used for searching for
+  bytes in arch native chunk sizes (e.g. 8 bytes on 64-bit).